# crates/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "easyImmerse API",
    "description": "The REST API of easyImmerse.",
    "license": {
      "name": "AGPL-3.0-only",
      "identifier": "AGPL-3.0-only"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/dictionaries": {
      "get": {
        "tags": [
          "dictionaries"
        ],
        "operationId": "listDictionaries",
        "responses": {
          "200": {
            "description": "Every imported dictionary, in import order",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListDictionariesResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "421": {
            "description": "Unexpected Host header",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_token": []
          }
        ]
      },
      "post": {
        "tags": [
          "dictionaries"
        ],
        "summary": "Starts importing the file and answers with the job to poll at `/dictionaries/imports/{id}`,\nsince a large dictionary takes longer to store than a browser waits for a response.",
        "operationId": "importDictionary",
        "parameters": [
          {
            "name": "fileName",
            "in": "query",
            "description": "The name of the uploaded file, whose extension tells formats such as MDict and CSV apart.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "columns",
            "in": "query",
            "description": "What each column of a table holds, as column roles separated by commas, in place of the detected layout.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "hasHeader",
            "in": "query",
            "description": "Whether the first row of a table is a header. Read only together with `columns`; false when left out.",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "description": "One dictionary file: a zip or tar archive, an MDict `.mdx`, a CSV or TSV file, and so on",
          "content": {
            "application/octet-stream": {}
          }
        },
        "responses": {
          "202": {
            "description": "The import was started",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportJobStarted"
                }
              }
            }
          },
          "400": {
            "description": "The chosen columns are not column roles",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "421": {
            "description": "Unexpected Host header",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_token": []
          }
        ]
      }
    },
...
    "/dictionaries/imports/{id}": {
      "get": {
        "tags": [
          "dictionaries"
        ],
        "operationId": "getImportJob",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The import job id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Where the import stands",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportJobStatus"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "No import job has the id, or its outcome has been forgotten",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "421": {
            "description": "Unexpected Host header",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_token": []
          }
        ]
      }
    },
...
    "/dictionaries/preview": {
      "post": {
        "tags": [
          "dictionaries"
        ],
        "summary": "Detects what each column of a table holds and returns that layout with the table's first rows,\nso that the user can check it before importing.",
        "operationId": "previewDictionaryTable",
        "parameters": [
          {
            "name": "fileName",
            "in": "query",
            "description": "The name of the uploaded file, whose extension marks it as a CSV, TSV or Tabfile table.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "A CSV, TSV or Tabfile table, or an archive holding one",
          "content": {
            "application/octet-stream": {}
          }
        },
        "responses": {
          "200": {
            "description": "The detected layout and the first rows",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TablePreview"
                }
              }
            }
          },
          "400": {
            "description": "The file could not be read as a table",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "421": {
            "description": "Unexpected Host header",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_token": []
          }
        ]
      }
    },
...
    "/dictionaries/{id}": {
      "delete": {
        "tags": [
          "dictionaries"
        ],
        "operationId": "deleteDictionary",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The dictionary id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The dictionary and everything it stored were deleted"
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "No dictionary has the id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "421": {
            "description": "Unexpected Host header",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_token": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "description": "The body of every error response.",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "A stable, machine-readable identifier such as `unauthorized`."
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "ColumnRole": {
        "type": "string",
        "description": "What a column of a table holds, as a person choosing the columns of a table sees it.",
        "enum": [
          "term",
          "reading",
          "definition",
          "alternates",
          "tags",
          "frequency",
          "ignored"
        ]
      },
...
      "DictionaryFormatKind": {
        "type": "string",
        "enum": [
          "yomitan",
          "stardict",
          "mdict",
          "csv"
        ]
      },
...
      "DictionarySummary": {
        "type": "object",
        "required": [
          "id",
          "title",
          "format",
          "entry_count",
          "term_meta_count",
          "tag_count",
          "kanji_count",
          "kanji_meta_count",
          "media_count"
        ],
        "properties": {
          "entry_count": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "format": {
            "$ref": "#/components/schemas/DictionaryFormatKind"
          },
          "id": {
            "type": "string"
          },
          "kanji_count": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "kanji_meta_count": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "media_count": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "source_language": {
            "type": [
              "string",
              "null"
            ],
            "description": "The language of the words looked up, as a BCP 47 tag, when the dictionary states it."
          },
          "tag_count": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "target_language": {
            "type": [
              "string",
              "null"
            ],
            "description": "The language of the definitions, as a BCP 47 tag, when the dictionary states it."
          },
          "term_meta_count": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "title": {
            "type": "string"
          }
        }
      },
...
      "ImportJobStarted": {
        "type": "object",
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "type": "string"
          }
        }
      },
      "ImportJobState": {
        "type": "string",
        "enum": [
          "running",
          "done",
          "failed"
        ]
      },
      "ImportJobStatus": {
        "type": "object",
        "required": [
          "state",
          "progress"
        ],
        "properties": {
          "dictionary": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/DictionarySummary",
                "description": "The stored dictionary, once the job is done."
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/ApiError",
                "description": "Why the import failed, once the job has failed."
              },
              {
                "type": "null"
              }
            ]
          },
          "progress": {
            "$ref": "#/components/schemas/ImportProgress"
          },
          "state": {
            "$ref": "#/components/schemas/ImportJobState"
          }
        }
      },
...
      "ImportProgress": {
        "type": "object",
        "description": "How many items of each kind the import has stored so far.",
        "required": [
          "entries",
          "term_meta",
          "kanji",
          "kanji_meta",
          "tags",
          "media"
        ],
        "properties": {
          "entries": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "kanji": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "kanji_meta": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "media": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "tags": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "term_meta": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
...
      "ListDictionariesResponse": {
        "type": "object",
        "required": [
          "dictionaries"
        ],
        "properties": {
          "dictionaries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DictionarySummary"
            }
          }
        }
      },
...
      "TableLayout": {
        "type": "object",
        "description": "What each column of a table holds, and whether its first row is a header.",
        "required": [
          "columns",
          "hasHeader"
        ],
        "properties": {
          "columns": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ColumnRole"
            }
          },
          "hasHeader": {
            "type": "boolean"
          }
        }
      },
      "TablePreview": {
        "type": "object",
        "description": "The layout detected in a table, with its first rows, for a person to check before importing.",
        "required": [
          "layout",
          "rows"
        ],
        "properties": {
          "layout": {
            "$ref": "#/components/schemas/TableLayout"
          },
          "rows": {
            "type": "array",
            "items": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "description": "The first rows of the table, including any header row, split into cells."
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_token": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}